DETECTOR: Modify getFileProperties to cache md5 results
